Python/requirements.txt
# Azure Speech SDK
azure-cognitiveservices-speech==1.38.0

# Azure Storage (for batch transcription)
azure-storage-blob==12.19.0
azure-identity==1.16.0

# Web Framework
Flask==3.0.3
Flask-CORS==4.0.1

# Security
Flask-WTF==1.2.1
Flask-Limiter==3.5.0

# Caching
Flask-Caching==2.1.0

# Document generation
python-docx==1.1.2

# HTTP Client
requests==2.32.3
aiohttp==3.9.5  # Async HTTP client for parallel API calls (PERFORMANCE)

# Configuration
python-dotenv==1.0.1

# Utilities
pytz==2024.1

# File validation
python-magic-bin==0.4.14; platform_system == "Windows"
python-magic==0.4.27; platform_system != "Windows"
